+++ server/src/test/java/com/continuuity/loom/http/LoomAdminHandlerTest.java
@@ -22,6 +22,7 @@ import com.continuuity.loom.admin.ImageType;
 import com.continuuity.loom.admin.Provider;
 import com.continuuity.loom.admin.Service;
 import com.continuuity.loom.admin.Tenant;
+import com.continuuity.loom.admin.TenantSpecification;
 import com.continuuity.loom.common.conf.Constants;
 import com.continuuity.loom.common.queue.Element;
 import com.continuuity.loom.common.queue.QueueMetrics;
@@ -30,12 +31,12 @@ import com.google.common.base.Charsets;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
-import com.google.common.collect.Sets;
 import com.google.gson.Gson;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.reflect.TypeToken;
+import org.apache.http.Header;
 import org.apache.http.HttpResponse;
 import org.apache.http.message.BasicHeader;
 import org.jboss.netty.handler.codec.http.HttpResponseStatus;
@@ -47,6 +48,7 @@ import java.io.Reader;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.UUID;
 
 /**
  *
@@ -95,33 +97,36 @@ public class LoomAdminHandlerTest extends LoomServiceTestBase {
 
   @Test
   public void testForbiddenIfNonadminGetsQueueMetrics() throws Exception {
-    tenantStore.writeTenant(new Tenant(USER1_ACCOUNT.getTenantId(), USER1_ACCOUNT.getTenantId(), 10, 10, 100));
+    tenantStore.writeTenant(
+      new Tenant(UUID.randomUUID().toString(), new TenantSpecification(USER1_ACCOUNT.getTenantId(), 10, 10, 100)));
     assertResponseStatus(doGet("/v1/loom/metrics/queues", USER1_HEADERS), HttpResponseStatus.FORBIDDEN);
   }
 
   @Test
   public void testGetQueueMetrics() throws Exception {
     try {
-      tenantStore.writeTenant(new Tenant("tenantA", "tenantA", 10, 10, 100));
-      tenantStore.writeTenant(new Tenant("tenantB", "tenantB", 10, 10, 100));
-      tenantStore.writeTenant(new Tenant("tenantC", "tenantC", 10, 10, 100));
-      tenantStore.writeTenant(new Tenant("tenantD", "tenantD", 10, 10, 100));
-      provisionerQueues.add("tenantA", new Element("task1"));
-      provisionerQueues.add("tenantA", new Element("task2"));
-      provisionerQueues.add("tenantA", new Element("task3"));
-      provisionerQueues.take("tenantA", "consumer");
-      provisionerQueues.add("tenantB", new Element("task4"));
-      provisionerQueues.add("tenantC", new Element("task5"));
-      provisionerQueues.take("tenantC", "consumer");
-      provisionerQueues.add("tenantD", new Element("task6"));
-      provisionerQueues.add("tenantD", new Element("task7"));
-      provisionerQueues.take("tenantD", "consumer");
+      tenantStore.writeTenant(new Tenant("idA", new TenantSpecification("tenantA", 10, 10, 100)));
+      tenantStore.writeTenant(new Tenant("idB", new TenantSpecification("tenantB", 10, 10, 100)));
+      tenantStore.writeTenant(new Tenant("idC", new TenantSpecification("tenantC", 10, 10, 100)));
+      tenantStore.writeTenant(new Tenant("idD", new TenantSpecification("tenantD", 10, 10, 100)));
+      provisionerQueues.add("idA", new Element("task1"));
+      provisionerQueues.add("idA", new Element("task2"));
+      provisionerQueues.add("idA", new Element("task3"));
+      provisionerQueues.take("idA", "consumer");
+      provisionerQueues.add("idB", new Element("task4"));
+      provisionerQueues.add("idC", new Element("task5"));
+      provisionerQueues.take("idC", "consumer");
+      provisionerQueues.add("idD", new Element("task6"));
+      provisionerQueues.add("idD", new Element("task7"));
+      provisionerQueues.take("idD", "consumer");
 
       Map<String, QueueMetrics> expected = Maps.newHashMap();
       expected.put("tenantA", new QueueMetrics(2, 1));
       expected.put("tenantB", new QueueMetrics(1, 0));
       expected.put("tenantC", new QueueMetrics(0, 1));
       expected.put("tenantD", new QueueMetrics(1, 1));
+      expected.put(TENANT, new QueueMetrics(0, 0));
+      expected.put(Constants.SUPERADMIN_TENANT, new QueueMetrics(0, 0));
 
       assertQueueMetrics(Constants.SUPERADMIN_TENANT, expected);
       assertQueueMetrics("tenantA", ImmutableMap.of("tenantA", new QueueMetrics(2, 1)));
@@ -130,10 +135,10 @@ public class LoomAdminHandlerTest extends LoomServiceTestBase {
       assertQueueMetrics("tenantD", ImmutableMap.of("tenantD", new QueueMetrics(1, 1)));
     } finally {
       provisionerQueues.removeAll();
-      tenantStore.deleteTenant("tenantA");
-      tenantStore.deleteTenant("tenantB");
-      tenantStore.deleteTenant("tenantC");
-      tenantStore.deleteTenant("tenantD");
+      tenantStore.deleteTenantByName("tenantA");
+      tenantStore.deleteTenantByName("tenantB");
+      tenantStore.deleteTenantByName("tenantC");
+      tenantStore.deleteTenantByName("tenantD");
     }
   }